Moved Socket into its own file, minor refactors
